feat(runtime-vapor): simplify implement
